(test) refactored server test to pass every test and break down into module
