+++ test/server/index.js
@@ -1,0 +1,4 @@
+describe('', function() {
+  require('./ControllerSpec.js');
+  require('./HandlerSpec.js');
+});
